Only dispatch when event target and listener match
By default custom events will only be triggered when the listener and
the event target match so as to 'disable' keybinding when child elements
have focus. This will make typing into text inputs and other content
editable elements work as expected. Specify :always-active T or NIL to
explicitly turn this behavior on or off, either for all keybindings or
on an individual basis.