Revamp set-on-keys syntax & trigger conditions
The new syntax does away with some of the keyword arguments and
structures the keybinding definition so as to allow for an implicit
lambda to wrap the body of the handler. It introduces the :trigger
keyword argument to explicitly change the trigger condition.

Initially, no trigger condition was implemented which meant that
listeners were triggering handlers freely as the keys matched with
propagated events. This was a problem because it interfered with the
behavior of elements such as text inputs.

A default condition for triggering handlers when keys were pressed was
introduced to ensure that the event target and listener matched, but
this was found to be slightly too restrictive. This did enable text
inputs to behave as expected, but didn't allow keybindings to have scope
or be 'inherited'.

The user may want to define some global keybindings, but also they may
want certain keybindings to become active when certain parts of the
document have 'focus'. This requires some form of scoping and
inheritance. One, albeit hacky, way to achieve scoping is by setting
intermediate child DIVs with a tabIndex of -1 so that they can receive
keyboard events. That way when they have focus, and events arrive via
propagation, they have the chance to trigger the appropriate handlers.
Unfortunately enforcing the condition that event target and listener
must match doesn't allow for the global keybindings to operate since it
is the intermediate DIV that has the focus.

This commit opens up the default condition so that key matched events
freely trigger handers again, but only if the activeElement isn't an
input element, with the exception of the event target also being the
listener (so that if the input element is the listener it doesn't
exclude itself).